+++ src/components/DataBrowserHeaderBar/DataBrowserHeaderBar.react.js
@@ -16,7 +16,7 @@ export default
 @DragDropContext(HTML5Backend)
 class DataBrowserHeaderBar extends React.Component {
   render() {
-    let { headers, onResize, selectAll, onAddColumn, updateOrdering, readonly } = this.props;
+    let { headers, onResize, selectAll, onAddColumn, updateOrdering, readonly, preventSchemaEdits } = this.props;
     let elements = [
       // Note: bulk checkbox is disabled as all rows are selected (not just visible ones due to current lazy loading implementation)
       // TODO: add bulk checking only visible rows
@@ -61,8 +61,9 @@ class DataBrowserHeaderBar extends React.Component {
     if (headers.length % 2) {
       finalStyle.background = 'rgba(224,224,234,0.10)';
     }
+
     elements.push(
-      readonly ? null : (
+      readonly || preventSchemaEdits ? null : (
         <div key='add' className={styles.addColumn} style={finalStyle}>
           <a
             href='javascript:;'