+++ src/dashboard/Data/Browser/BrowserToolbar.react.js
@@ -33,6 +33,7 @@ let BrowserToolbar = ({
   onAddClass,
   onAttachRows,
   onAttachSelectedRows,
+  onCloneSelectedRows,
   onExport,
   onRemoveColumn,
   onDeleteRows,
@@ -40,12 +41,14 @@ let BrowserToolbar = ({
   onChangeCLP,
   onRefresh,
   hidePerms,
+  isUnique,
 
   enableDeleteAllRows,
   enableExportClass,
   enableSecurityDialog,
 
-  isUnique,
+  enableColumnManipulation,
+  enableClassManipulation,
 }) => {
   let selectionLength = Object.keys(selection).length;
   let details = [];
@@ -94,8 +97,8 @@ let BrowserToolbar = ({
     menu = (
       <BrowserMenu title='Edit' icon='edit-solid' disabled={isUnique}>
         <MenuItem text='Add a row' onClick={onAddRow} />
-        <MenuItem text='Add a column' onClick={onAddColumn} />
-        <MenuItem text='Add a class' onClick={onAddClass} />
+        {enableColumnManipulation ? <MenuItem text='Add a column' onClick={onAddColumn} /> : <noscript />}
+        {enableClassManipulation ? <MenuItem text='Add a class' onClick={onAddClass} /> : <noscript />}
         <Separator />
         <MenuItem
           disabled={!selectionLength}
@@ -104,12 +107,18 @@ let BrowserToolbar = ({
         />
         <Separator />
         <MenuItem
+          disabled={!selectionLength}
+          text={`Clone ${selectionLength <= 1 ? 'this row' : 'these rows'}`}
+          onClick={onCloneSelectedRows}
+        />
+        <Separator />
+        <MenuItem
           disabled={selectionLength === 0}
           text={selectionLength === 1 && !selection['*'] ? 'Delete this row' : 'Delete these rows'}
           onClick={() => onDeleteRows(selection)} />
-        <MenuItem text='Delete a column' onClick={onRemoveColumn} />
+        {enableColumnManipulation ? <MenuItem text='Delete a column' onClick={onRemoveColumn} /> : <noscript />}
         {enableDeleteAllRows ? <MenuItem text='Delete all rows' onClick={() => onDeleteRows({ '*': true })} /> : <noscript />}
-        <MenuItem text='Delete this class' onClick={onDropClass} />
+        {enableClassManipulation ? <MenuItem text='Delete this class' onClick={onDropClass} /> : <noscript />}
         {enableExportClass ? <Separator /> : <noscript />}
         {enableExportClass ? <MenuItem text='Export this data' onClick={onExport} /> : <noscript />}
       </BrowserMenu>