+++ src/lib/ParseApp.js
@@ -42,6 +42,7 @@ export default class ParseApp {
     primaryBackgroundColor,
     secondaryBackgroundColor,
     supportedPushLocales,
+    preventSchemaEdits
   }) {
     this.name = appName;
     this.createdAt = created_at ? new Date(created_at) : new Date();
@@ -65,6 +66,7 @@ export default class ParseApp {
     this.primaryBackgroundColor=primaryBackgroundColor;
     this.secondaryBackgroundColor=secondaryBackgroundColor;
     this.supportedPushLocales = supportedPushLocales ? supportedPushLocales : [];
+    this.preventSchemaEdits = preventSchemaEdits || false;
 
     if(!supportedPushLocales) {
       console.warn(`Missing push locales for '` + appName + `', see this link for details on setting localizations up. https://github.com/parse-community/parse-dashboard#configuring-localized-push-notifications`);